컴포넌트별로 분리
