+++ src/components/RedirectPage/SharedStyles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const PageContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+`;
