+++ src/components/UserInfoPage/SharedStyles.ts
@@ -1,0 +1,61 @@
+import styled from "styled-components";
+
+export const PageContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  padding: 1rem;
+`;
+
+export const FormContainer = styled.div`
+  padding: 2rem;
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  background-color: #ffffff;
+  width: 100%;
+  max-width: 400px;
+`;
+
+export const Title = styled.h2`
+  margin-bottom: 1.5rem;
+  text-align: center;
+`;
+
+export const FormGroup = styled.div`
+  margin-bottom: 1rem;
+`;
+
+export const Label = styled.label`
+  display: block;
+  margin-bottom: 0.5rem;
+`;
+
+export const Input = styled.input`
+  width: 100%;
+  padding: 0.5rem;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+`;
+
+export const Select = styled.select`
+  width: 100%;
+  padding: 0.5rem;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+`;
+
+export const Button = styled.button`
+  width: 100%;
+  padding: 0.75rem;
+  border: none;
+  border-radius: 4px;
+  background-color: #007bff;
+  color: white;
+  cursor: pointer;
+
+  &:hover {
+    background-color: #0056b3;
+  }
+`;
