+++ src/pages/TodaysCare.tsx
@@ -1,8 +1,10 @@
 import { useState, useEffect } from 'react';
-import styled from 'styled-components';
-import { ContentWrapper } from '../components/Layout';
 import { getUVIndex } from '../services/uv_indexApi';
 import type { UVIndexResponse } from '../services/types';
+
+import CareHeader from '../components/TodaysCare/CareHeader';
+import UvIndexSection from '../components/TodaysCare/UvIndexSection';
+import CareTipsSection from '../components/TodaysCare/CareTipsSection';
 
 // --- 타입 정의 --- //
 interface Tip {
@@ -25,215 +27,6 @@ interface CareLevelData {
 interface CareData {
   [key: string]: CareLevelData;
 }
-
-const LocationStatus = styled.p<{ $isRealTime: boolean }>`
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  margin-bottom: 0.25rem;
-  font-weight: ${props => props.$isRealTime ? '600' : '400'};
-  color: ${props => props.$isRealTime ? '#1E293B' : '#64748B'};
-  font-size: ${props => props.$isRealTime ? '1.125rem' : '1rem'};
-
-  &::before {
-    content: '●';
-    color: ${props => props.$isRealTime ? '#22c55e' : '#94a3b8'};
-    font-size: 0.875rem;
-    margin-right: 0.25rem;
-  }
-`;
-
-
-// --- 스타일 컴포넌트 정의 --- //
-const Header = styled.header`
-  text-align: center;
-  padding: 3rem 0;
-
-  h1 {
-    font-size: clamp(2rem, 5vw, 2.5rem); 
-    font-weight: 700;
-    color: #1E293B;
-    margin: 0;
-  }
-  p {
-    font-size: clamp(1rem, 2vw, 1.25rem); 
-    color: #64748B;
-    margin-top: 0.5rem;
-  }
-`;
-
-const UvInfoBox = styled.div`
-  width: 100%;
-  background-color: #F0F9FF;
-  padding: 2.5rem 0;
-  box-sizing: border-box;
-
-  @media (max-width: 768px) {
-    padding: 2rem 0;
-  }
-`;
-
-const UvInfoInnerWrapper = styled(ContentWrapper)`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-    gap: 2rem;
-    text-align: center;
-  }
-`;
-
-const UvIndexDisplay = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-  flex-shrink: 0;
-
-  @media (max-width: 480px) {
-    flex-direction: column;
-    gap: 1rem;
-  }
-`;
-
-const UvIndexVisual = styled.div<{ $bgColor: string }>`
-  width: 8rem; 
-  height: 8rem; 
-  border-radius: 9999px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  color: white;
-  box-shadow: 0 0.625rem 0.9375rem -0.1875rem rgba(0, 0, 0, 0.1), 0 0.25rem 0.375rem -0.125rem rgba(0, 0, 0, 0.05);
-  background-color: ${props => props.$bgColor};
-  transition: background-color 0.3s ease;
-
-  span:first-child {
-    font-size: 1rem; 
-    font-weight: 500;
-  }
-  span:last-child {
-    font-size: clamp(2.5rem, 6vw, 3rem); 
-    font-weight: 700;
-  }
-`;
-
-const UvIndexText = styled.div`
-  p {
-    color: #64748B;
-    margin: 0;
-    font-size: 1rem;
-  }
-  h2 {
-    font-size: clamp(1.8rem, 4vw, 2.25rem); 
-    font-weight: 700;
-    color: #1E293B;
-    margin: 0.25rem 0 0 0;
-  }
-`;
-
-const UvSummary = styled.div`
-  text-align: right;
-  flex-basis: 50%;
-
-  p {
-    font-size: clamp(0.9rem, 1.5vw, 1.3rem);
-    word-break: keep-all;
-  }
-  p:first-child {
-    font-weight: 600;
-    color: #334155;
-    font-size: clamp(1rem, 2vw, 1.5rem);
-  }
-  p:last-child {
-    color: #64748B;
-    margin-top: 0.5rem;
-    line-height: 1.6;
-  }
-
-  @media (max-width: 768px) {
-    text-align: center;
-    flex-basis: 100%;
-  }
-`;
-
-const CareTipsContainer = styled.div`
-  width: 100%;
-  padding: 2.5rem 0 4rem 0;
-`;
-
-const TabsContainer = styled.div`
-  display: flex;
-  border-bottom: 0.0625rem solid #E2E8F0;
-  overflow-x: auto;
-  &::-webkit-scrollbar { height: 0.25rem; }
-  &::-webkit-scrollbar-thumb { background: #CBD5E1; border-radius: 0.625rem; }
-`;
-
-const TabButton = styled.button<{ $active: boolean }>`
-  padding: 0.75rem 1.25rem;
-  border: none;
-  background-color: transparent;
-  color: ${props => props.$active ? '#0891B2' : '#64748B'};
-  font-weight: ${props => props.$active ? '600' : '500'};
-  cursor: pointer;
-  transition: color 0.2s, border-color 0.2s;
-  border-bottom: 0.1875rem solid ${props => props.$active ? '#0891B2' : 'transparent'};
-  flex-shrink: 0;
-  white-space: nowrap;
-  font-size: clamp(0.9rem, 1.8vw, 1.2rem); 
-  display: flex;
-  align-items: center;
-  gap: 0.3rem;
-`;
-
-const TabRange = styled.span`
-  font-size: 0.8rem;
-  color: #94A3B8;
-  font-weight: 400;
-`;
-
-const TipsContent = styled.div`
-  padding-top: 2rem;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(17.5rem, 1fr));
-  gap: 1.5rem;
-`;
-
-const TipCard = styled.div`
-  display: flex;
-  align-items: flex-start;
-  padding: 1.5rem;
-  background-color: white;
-  border: 0.0625rem solid #E2E8F0;
-  border-radius: 0.75rem;
-  box-shadow: 0 0.0625rem 0.125rem 0 rgba(0, 0, 0, 0.05);
-  transition: box-shadow 0.2s, border-color 0.2s;
-
-  &:hover {
-    border-color: #A5F3FC;
-    box-shadow: 0 0.25rem 0.375rem -0.0625rem rgba(0, 0, 0, 0.1), 0 0.125rem 0.25rem -0.0625rem rgba(0, 0, 0, 0.06);
-  }
-
-  .icon {
-    font-size: 1.875rem;
-    margin-right: 1rem;
-  }
-  h4 {
-    font-weight: 700;
-    color: #334155;
-    margin: 0;
-    font-size: clamp(1rem, 2vw, 1.3rem);
-  }
-  p {
-    color: #64748B;
-    margin-top: 0.25rem;
-    text-align: left;
-    font-size: clamp(0.85rem, 1.5vw, 1.05rem);
-  }
-`;
 
 // --- 데이터 --- //
 const careData: CareData = {
@@ -265,7 +58,6 @@ const getUVIndexFromCareLevel = (careLevel: string): number => {
   }
 };
 
-// --- 컴포넌트 --- //
 function TodaysCare() {
   const [activeTab, setActiveTab] = useState<string>('보통');
   const [uvData, setUvData] = useState<UVIndexResponse | null>(null);
@@ -273,7 +65,6 @@ function TodaysCare() {
   const [error, setError] = useState<string | null>(null);
   const [isRealTimeData, setIsRealTimeData] = useState<boolean>(true);
 
-  // 페이지 로드 시 UV 지수 API 호출
   useEffect(() => {
     const fetchUVIndex = async () => {
       try {
@@ -289,12 +80,12 @@ function TodaysCare() {
           setIsRealTimeData(true);
         } else {
           setError('UV 지수 데이터를 찾을 수 없습니다.');
-          setIsRealTimeData(false); // 데이터 없으면 실시간 아님
+          setIsRealTimeData(false);
         }
       } catch (err) {
         console.error('UV 지수를 불러오는데 실패했습니다:', err);
         setError('UV 지수를 불러오는데 실패했습니다.');
-        setIsRealTimeData(false); // 에러 시 실시간 아님
+        setIsRealTimeData(false);
       } finally {
         setIsLoading(false);
       }
@@ -303,13 +94,11 @@ function TodaysCare() {
     fetchUVIndex();
   }, []);
 
-  // 탭 클릭 핸들러
   const handleTabClick = (level: string) => {
     setActiveTab(level);
-    setIsRealTimeData(false); // 탭을 클릭하면 더 이상 실시간 데이터가 아님
+    setIsRealTimeData(false);
   };
 
-  // 실시간 데이터로 돌아가기 버튼 핸들러
   const handleRealTimeClick = () => {
     if (uvData && uvData.now) {
       const uvIndex = parseInt(uvData.now, 10);
@@ -319,106 +108,34 @@ function TodaysCare() {
     }
   };
 
-  // 현재 탭에 맞는 데이터 선택
   const currentCareData = careData[activeTab];
 
-  // UV 지수 표시 값 결정
   const displayUVIndex = isRealTimeData && uvData
     ? uvData.now
     : getUVIndexFromCareLevel(activeTab).toString();
 
-  // 로딩 중일 때 UI
-  if (isLoading) {
-    return (
-      <>
-        <ContentWrapper>
-          <Header>
-            <h1>오늘의 맞춤 케어</h1>
-            <p>현재 자외선 지수를 확인하고, 내 피부를 위한 팁을 알아보세요.</p>
-          </Header>
-        </ContentWrapper>
-        <UvInfoBox>
-          <UvInfoInnerWrapper>
-            <div style={{ textAlign: 'center', padding: '2rem' }}>
-              <p>자외선 지수를 불러오는 중...</p>
-            </div>
-          </UvInfoInnerWrapper>
-        </UvInfoBox>
-      </>
-    );
-  }
-
   return (
     <>
-      <ContentWrapper>
-        <Header>
-          <h1>오늘의 맞춤 케어</h1>
-          <p>현재 자외선 지수를 확인하고, 내 피부를 위한 팁을 알아보세요.</p>
-        </Header>
-      </ContentWrapper>
+      <CareHeader />
       
       <main>
-        <UvInfoBox>
-          <UvInfoInnerWrapper>
-            <UvIndexDisplay>
-              <UvIndexVisual $bgColor={currentCareData.color}>
-                <span>UV 지수</span>
-                <span>{displayUVIndex}</span>
-              </UvIndexVisual>
-              <UvIndexText>
-                <LocationStatus $isRealTime={isRealTimeData}>
-                  {isRealTimeData ? `${uvData?.location || '현재 위치'} (실시간)` : '케어 단계별 정보'}
-                </LocationStatus>
-                <h2>{activeTab}</h2>
-                {isRealTimeData && uvData?.date && (
-                  <p style={{ fontSize: '0.875rem', color: '#64748B', margin: '0.25rem 0 0 0' }}>
-                    측정 시간: {uvData.date}
-                  </p>
-                )}
-                {error && <p style={{ color: '#EF4444', fontSize: '0.875rem' }}>{error}</p>}
-              </UvIndexText>
-            </UvIndexDisplay>
-            <UvSummary>
-              <p><strong>{currentCareData.summary.title}</strong></p>
-              <p>{currentCareData.summary.description}</p>
-            </UvSummary>
-          </UvInfoInnerWrapper>
-        </UvInfoBox>
+        <UvIndexSection
+          uvData={uvData}
+          isLoading={isLoading}
+          error={error}
+          isRealTimeData={isRealTimeData}
+          currentCareData={currentCareData}
+          displayUVIndex={displayUVIndex}
+        />
 
-        <ContentWrapper>
-          <CareTipsContainer>
-            <TabsContainer>
-              <TabButton
-                $active={isRealTimeData}
-                onClick={handleRealTimeClick}
-                disabled={!uvData}
-              >
-                🌡️ 실시간
-              </TabButton>
-              {Object.entries(careData).map(([level, data]) => (
-                <TabButton
-                  key={level}
-                  $active={!isRealTimeData && activeTab === level}
-                  onClick={() => handleTabClick(level)}
-                >
-                  {level}
-                  <TabRange>({data.range})</TabRange>
-                </TabButton>
-              ))}
-            </TabsContainer>
-            <TipsContent>
-              {currentCareData.tips.map((tip, index) => (
-                <TipCard key={index}>
-                  <span className="icon" role="img" aria-label="icon">{tip.icon}</span>
-                  <div>
-                    <h4>{tip.title}</h4>
-                    <p>{tip.description}</p>
-                  </div>
-                </TipCard>
-              ))}
-            </TipsContent>
-          </CareTipsContainer>
-        </ContentWrapper>
+        <CareTipsSection
+          careData={careData}
+          activeTab={activeTab}
+          isRealTimeData={isRealTimeData}
+          uvData={uvData}
+          handleRealTimeClick={handleRealTimeClick}
+          handleTabClick={handleTabClick}
+        />
       </main>
     </>
   );
