+++ src/services/dashboardApi.ts
@@ -1,4 +1,5 @@
 import apiClient from "./apiClient";
+import type { SkinTypeScore } from "./types";
 
 export interface DiagnosisRecord {
   id: number;
@@ -14,7 +15,7 @@ export interface DiagnosisRecord {
 }
 
 export interface DashboardData {
-  recent_skinType_scores: string[]; // 실제 타입에 따라 더 구체화 필요
+  recent_skinType_scores: SkinTypeScore[];
   recent_diagnosis_records: DiagnosisRecord[];
   // 백엔드 응답에 없는 필드들은 일단 제거하거나 optional로 처리합니다.
   // chartLabels?: string[];
