+++ src/services/types.ts
@@ -103,4 +103,27 @@ export interface ApiError {
   success: false;
   error: string;
   details?: unknown;
-} 
+}
+
+export interface SkinTypeScore {
+  date: string;
+  score: number;
+}
+
+export interface DiagnosisRecord {
+  id: number;
+  user_id: number;
+  disease_name: string | null;
+  skin_type_id: number | null;
+  confidence: number;
+  image: string;
+  after: string | null;
+  created_at: string;
+  updated_at: string;
+  is_deleted: boolean;
+}
+
+export interface DashboardData {
+  recent_skinType_scores: SkinTypeScore[];
+  recent_diagnosis_records: DiagnosisRecord[];
+}
